TST/REF: collect/de-dup index tests (#43914)
