+++ pandas/tests/indexes/categorical/test_reindex.py
@@ -1,13 +1,10 @@
 import numpy as np
-import pytest
 
 from pandas import (
     Categorical,
     CategoricalIndex,
-    DataFrame,
     Index,
     Interval,
-    Series,
 )
 import pandas._testing as tm
 
@@ -65,45 +62,6 @@ class TestReindex:
         res, indexer = c.reindex(["a", "b"])
         tm.assert_index_equal(res, Index(["a", "b"]), exact=True)
         tm.assert_numpy_array_equal(indexer, np.array([-1, -1], dtype=np.intp))
-
-    def test_reindex_missing_category(self):
-        # GH: 18185
-        ser = Series([1, 2, 3, 1], dtype="category")
-        msg = r"Cannot setitem on a Categorical with a new category \(-1\)"
-        with pytest.raises(TypeError, match=msg):
-            ser.reindex([1, 2, 3, 4, 5], fill_value=-1)
-
-    @pytest.mark.parametrize(
-        "index_df,index_res,index_exp",
-        [
-            (
-                CategoricalIndex([], categories=["A"]),
-                Index(["A"]),
-                Index(["A"]),
-            ),
-            (
-                CategoricalIndex([], categories=["A"]),
-                Index(["B"]),
-                Index(["B"]),
-            ),
-            (
-                CategoricalIndex([], categories=["A"]),
-                CategoricalIndex(["A"]),
-                CategoricalIndex(["A"]),
-            ),
-            (
-                CategoricalIndex([], categories=["A"]),
-                CategoricalIndex(["B"]),
-                CategoricalIndex(["B"]),
-            ),
-        ],
-    )
-    def test_reindex_not_category(self, index_df, index_res, index_exp):
-        # GH: 28690
-        df = DataFrame(index=index_df)
-        result = df.reindex(index=index_res)
-        expected = DataFrame(index=index_exp)
-        tm.assert_frame_equal(result, expected)
 
     def test_reindex_categorical_added_category(self):
         # GH 42424
