+++ pandas/tests/indexes/datetimes/test_date_range.py
@@ -121,6 +121,41 @@ class TestTimestampEquivDateRange:
 
 
 class TestDateRanges:
+    @pytest.mark.parametrize("freq", ["N", "U", "L", "T", "S", "H", "D"])
+    def test_date_range_edges(self, freq):
+        # GH#13672
+        td = Timedelta(f"1{freq}")
+        ts = Timestamp("1970-01-01")
+
+        idx = date_range(
+            start=ts + td,
+            end=ts + 4 * td,
+            freq=freq,
+        )
+        exp = DatetimeIndex(
+            [ts + n * td for n in range(1, 5)],
+            freq=freq,
+        )
+        tm.assert_index_equal(idx, exp)
+
+        # start after end
+        idx = date_range(
+            start=ts + 4 * td,
+            end=ts + td,
+            freq=freq,
+        )
+        exp = DatetimeIndex([], freq=freq)
+        tm.assert_index_equal(idx, exp)
+
+        # start matches end
+        idx = date_range(
+            start=ts + td,
+            end=ts + td,
+            freq=freq,
+        )
+        exp = DatetimeIndex([ts + td], freq=freq)
+        tm.assert_index_equal(idx, exp)
+
     def test_date_range_near_implementation_bound(self):
         # GH#???
         freq = Timedelta(1)
@@ -717,7 +752,7 @@ class TestDateRanges:
         result = date_range(start, periods=2, tz="US/Eastern")
         assert len(result) == 2
 
-    def test_timezone_comparaison_assert(self):
+    def test_timezone_comparison_assert(self):
         start = Timestamp("20130220 10:00", tz="US/Eastern")
         msg = "Inferred time zone not equal to passed time zone"
         with pytest.raises(AssertionError, match=msg):
