+++ pandas/tests/indexes/period/test_period.py
@@ -245,25 +245,6 @@ class TestPeriodIndex(DatetimeLike):
         assert not index.is_(index - 2)
         assert not index.is_(index - 0)
 
-    def test_index_duplicate_periods(self):
-        # monotonic
-        idx = PeriodIndex([2000, 2007, 2007, 2009, 2009], freq="A-JUN")
-        ts = Series(np.random.randn(len(idx)), index=idx)
-
-        result = ts["2007"]
-        expected = ts[1:3]
-        tm.assert_series_equal(result, expected)
-        result[:] = 1
-        assert (ts[1:3] == 1).all()
-
-        # not monotonic
-        idx = PeriodIndex([2000, 2007, 2007, 2009, 2007], freq="A-JUN")
-        ts = Series(np.random.randn(len(idx)), index=idx)
-
-        result = ts["2007"]
-        expected = ts[idx == "2007"]
-        tm.assert_series_equal(result, expected)
-
     def test_index_unique(self):
         idx = PeriodIndex([2000, 2007, 2007, 2009, 2009], freq="A-JUN")
         expected = PeriodIndex([2000, 2007, 2009], freq="A-JUN")
@@ -291,12 +272,6 @@ class TestPeriodIndex(DatetimeLike):
         tm.assert_index_equal(idx.year, exp)
         exp = Index([1, 2, -1, 3, 4], dtype=np.int64, name="name")
         tm.assert_index_equal(idx.month, exp)
-
-    def test_pindex_qaccess(self):
-        pi = PeriodIndex(["2Q05", "3Q05", "4Q05", "1Q06", "2Q06"], freq="Q")
-        s = Series(np.random.rand(len(pi)), index=pi).cumsum()
-        # Todo: fix these accessors!
-        assert s["05Q4"] == s[2]
 
     def test_pindex_multiples(self):
         expected = PeriodIndex(
